+++ src/backend/config/KsnipConfigOptions.h
@@ -41,6 +41,7 @@ public:
 	static QString useTrayIconString();
 	static QString minimizeToTrayString();
 	static QString closeToTrayString();
+	static QString startMinimizedToTray();
 	static QString textFontString();
 	static QString numberFontString();
 	static QString itemShadowEnabledString();